requirements.txt
fastapi==0.116.1
langchain==0.3.27
langchain_chroma==0.2.5
langchain_core==0.3.75
langchain_google_genai==2.1.10
langgraph==0.6.6
langsmith==0.4.19
loguru==0.7.3
pydantic==2.11.7
pydantic-settings==2.2.1
python-dotenv==1.1.1
requests==2.32.5
streamlit==1.48.0
unstructured[all-docs]==0.18.11
uvicorn==0.34.0
pandas==2.2.3
